[level 1] Title: 콜라츠 추측, Time: 0.09 ms, Memory: 88.6 MB -BaekjoonHub

+++ 프로그래머스/1/12943. 콜라츠 추측/콜라츠 추측.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int solution(long num) {
+    public int solution(double num) {
         int answer = 0;
         
         if (num == 1){
@@ -16,7 +16,7 @@ class Solution {
         }
         if (answer >= 500) {
             return -1;
-        }
+            }
     return answer;
     }
 }
